cherrypick for Ansible UI fixes PR
